Initial 'copilot for cli' clone. It does what I need.
Today I found copilot-cli[1], and was in awe. A problem I hit all the
time has now been solved through the help of AI. A tool to make
information retreival more seemless. No more forgetting the exact flags
for a command. With intention an a simple search interface, I can now
find what is needed.

Unfortunatley, this tool is not yet available for me. There is a
waitlist to be able to use it.

That's where this repository comes in.  The python script in this repo
attempts to provide similar functionality by leveraging gpt-3.5-turbo
language model. demo.txt demonstrates its use, and performs well enough
for me!

If you are waitlisted on copilot-cli, you may find this script useful...

[1] https://githubnext.com/projects/copilot-cli/